requirements.txt
# Harbor Tasks Verification Dependencies
# Pinned versions for reproducibility

# Core dependencies
Jinja2>=3.1.6
litellm>=1.77.4
pandas>=2.3.2
PyYAML>=6.0.3
python-dotenv>=1.1.1

# Async support
aiohttp>=3.8.0

# Data processing
numpy>=2.3.3
pyarrow>=21.0.0

# Utilities
tqdm>=4.67.1
requests>=2.32.5

